* Using persistent key value database for gateway - Chronicle map. * Handling of failed scan on the gateway side. * Fixed bug of incorrect displayed file name in report window in client-web. * Some refactoring & code cleanup.

+++ client-web/src/sk/csirt/viruschecker/client/web/config/DependencyInjectionConfig.kt
@@ -8,7 +8,7 @@ import sk.csirt.viruschecker.client.service.ReportByHashService
 import sk.csirt.viruschecker.client.web.parsedArgs
 
 val webClientDependencyInjectionModule = module {
-    single { httpClient(parsedArgs.socketTimeout.toInt()) }
+    single { httpClient(parsedArgs.socketTimeout) }
     single { DriverInfoGatewayService(parsedArgs.gateway, get()) }
     single { ReportByHashService(parsedArgs.gateway, get()) }
     single { MultiScanService(parsedArgs.gateway, get()) }
